refactor: extract shell control projection
